Implement content addressable storage
The decompression is extremely slow at the moment. This causes files to
be cached based on their hash, meaning that once we have the index, we
can make one local read call and if it succeeds, we can skip both
downloading and decompressing of bundle files, and just return the
pointed to file.

It makes a ton of sense for dynamic use, but maybe should be optional
for non-dynamic uses? That is, if the cli utility is being used to
extract, then we're by definition writing the file out, so saving it
twice is kinda silly. Maybe in that case we should hash the file if it
exists locally and only fetch it if the hash doesn't match?